requirement.txt
RPi.GPIO==0.6.3
gpiozero==1.3.1
numpy==1.8.2
pyserial==2.6
sense-emu==1.0
sense-hat==2.2.0

